CardList finish

+++ PokemonTCGManager/app/src/main/java/be/leocheikhboukal/pokemontcgmanager/ui/deck/DeckAddScreen.kt
@@ -62,7 +62,7 @@ object DeckAddDestination : NavigationDestination {
 fun DeckAddScreen(
     navigateBack: () -> Unit,
     onNavigateUp: () -> Unit,
-    navigateToCardSearch: () -> Unit,
+    navigateToCardSearch: (Int) -> Unit,
     navigateToDeck: (Int) -> Unit,
     navigateToUser: (Int) -> Unit,
     modifier: Modifier = Modifier,
